requirements.txt
Flask==3.0.0
boto3>=1.35.50
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
